Adding example implementation of recursive factorial computation returning all of the intermediate results..

+++ problems/factorial.cpp
@@ -7,7 +7,7 @@
 
 #include <iostream>
 
-float factorial (int number)
+int factorial (int number)
 {
     if (number > 1) {
         // Recursive case.
